Project7Datapipeline/DockerCompile/slackyContainer/requirements.txt
slackclient
sqlalchemy
datetime
psycopg2-binary
